Renaming MMPreferences to SuitePreferences
